refactor: split the sidebar golden's chip blocks out of the state matrix
The three chip blocks are a different kind of enumeration from everything
around them: each maps ONE persisted field plus a staleness bit to one cell,
while the rest of the matrix builds whole row views and crosses activity
against a dozen axes. They share only the `task()` fixture.

The golden file is byte-identical after the move, which is the proof the split
changed nothing; `sidebar-state-matrix.ts` was 527 lines against the 500 cap.
